% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash_links-data.R
\docType{data}
\name{splash_links}
\alias{splash_links}
\title{image links}
\format{
An object of class \code{"character"}
}
\usage{
splash_links
}
\description{
vector of image links used by package
}
\keyword{datasets}
